+++ ADC.h
@@ -19,13 +19,13 @@ void ADC_Setup(void);
  * Sets up the item to be sampled by the ADC
  * Passed x to define item
  */
-void ADC_DestinationSetup(int x);
+void ADC_DestinationSetup(uint8_t x);
 
 
 /*
  * Contains code to sample ADC - measuring light levels 
  * Passed x to specify item to sample
  */
-int ADC_Read(int x);
+int ADC_Read(uint8_t x);
 
 #endif //ADC_H
